[UPD] Code refactoring between res.partner and res.partner.backend
Move of several methods in res.partner.backend model
