+++ lcc_cyclos_base/models/__init__.py
@@ -1,4 +1,5 @@
 from . import res_config_settings
 from . import company
 from . import res_partner
+from . import res_partner_backend
 from . import account_invoice
