+++ lcc_cyclos_base/models/res_partner.py
@@ -8,69 +8,10 @@ import logging
 _logger = logging.getLogger(__name__)
 
 
-class ResPartnerBackend(models.Model):
-    """Add backend commom property for local currency"""
-
-    _inherit = "res.partner.backend"
-
-    type = fields.Selection(selection_add=[("cyclos", "Cyclos")])
-    cyclos_create_response = fields.Text(string="Cyclos create response")
-    cyclos_id = fields.Char(string="Cyclos id")
-    cyclos_status = fields.Char(string="Cyclos Status")
-
-
 class ResPartner(models.Model):
     """Inherits partner, adds Cyclos fields in the partner form, and functions"""
 
     _inherit = "res.partner"
-
-    cyclos_create_response = fields.Text(string="Cyclos create response")
-    cyclos_active = fields.Boolean(string="Cyclos OK")
-    cyclos_id = fields.Char(string="Cyclos id")
-    cyclos_status = fields.Char(string="Cyclos Status")
-
-    def _build_cyclos_error_message(self, e):
-        json_error = e.response.json()
-        msg = ""
-        if json_error.get("code") == "validation":
-            if json_error.get("propertyErrors"):
-                error = json_error.get("propertyErrors")
-            elif json_error.get("generalErrors"):
-                error = json_error.get("generalErrors")
-            msg = ["  - %s: %s" % (k, ", ".join(v)) for k, v in error.items()]
-        return msg
-
-    def _cyclos_rest_call(
-        self, method, entrypoint, data={}, api_login=False, api_password=False
-    ):
-        headers = {"Content-type": "application/json", "Accept": "text/plain"}
-        requests.packages.urllib3.disable_warnings()
-        if not api_login:
-            api_login = self.env.user.company_id.cyclos_server_login
-        if not api_password:
-            api_password = self.env.user.company_id.cyclos_server_password
-        api_url = "%s%s" % (self.env.user.company_id.cyclos_server_url, entrypoint)
-        res = requests.request(
-            method.lower(),
-            api_url,
-            auth=HTTPBasicAuth(api_login, api_password),
-            verify=False,
-            data=json.dumps(data),
-            headers=headers,
-        )
-        try:
-            res.raise_for_status()
-        except requests.exceptions.HTTPError as e:
-            _logger.debug(e.response.json())
-            if e.response.status_code == 422:
-                msg = self._build_cyclos_error_message(e)
-                if msg != "":
-                    raise ValueError(
-                        "Cyclos serveur complained about:\n%s" % "\n".join(msg),
-                        e.response,
-                    )
-            raise
-        return res
 
     @property
     def _cyclos_backend_id(self):
@@ -118,10 +59,11 @@ class ResPartner(models.Model):
         self.ensure_one()
         data = super(ResPartner, self)._update_auth_data(password)
         # Update cyclos password with odoo one from authenticate session
+        backend = self._cyclos_backend()
         backend_json_data = self._cyclos_backend_json_data()
         if self._cyclos_backend_id and backend_json_data:
-            self.forceCyclosPassword(password)
-            new_token = self.createCyclosUserToken(self.id, password)
+            backend.forceCyclosPassword(password)
+            new_token = backend.createCyclosUserToken(self.id, password)
             if new_token:
                 for ua in backend_json_data[0]["accounts"]:
                     ua["token"] = new_token
@@ -136,13 +78,11 @@ class ResPartner(models.Model):
 
     def _update_search_data(self, backend_keys):
         self.ensure_one()
-        # _logger.debug('SEARCH: backend_keys = %s' % backend_keys)
         backend_data = self._cyclos_backend()
         data = super(ResPartner, self)._update_search_data(backend_keys)
         for backend_key in backend_keys:
             if backend_key.startswith("cyclos:") and backend_data.cyclos_id:
                 data[backend_key] = [backend_data.cyclos_id]
-        # _logger.debug('SEARCH: data %s' % data)
         return data
 
     def domains_is_unvalidated_currency_backend(self):
@@ -159,7 +99,7 @@ class ResPartner(models.Model):
     def backends(self):
         self.ensure_one()
         backends = super(ResPartner, self).backends()
-        if self.cyclos_id:
+        if self._cyclos_backend().cyclos_id:
             cyclos_serveur_url = self.env.user.company_id.cyclos_server_url
             remove = ["https://", "http://", "/api"]
             for value in remove:
@@ -205,22 +145,14 @@ class ResPartner(models.Model):
     def action_credit_cyclos_account(self, amount):
         for record in self:
             backend_data = record._cyclos_backend()
-            parent_backend_data = record.parent_id._cyclos_backend()
-            data = {
-                "amount": amount,
-                "description": "Credited by %s" % record.company_id.name,
-                "subject": backend_data.cyclos_id
-                if backend_data.status == "active"
-                else parent_backend_data.cyclos_id,
-                "type": "debit.toPro" if record.is_company else "debit.toUser",
-            }
-            _logger.debug("data: %s" % data)
-            res = record._cyclos_rest_call("POST", "/system/payments", data=data)
-            _logger.debug("res: %s" % res)
+            if backend_data.status != "active":
+                backend_data = record.parent_id._cyclos_backend()
+            backend_data.credit_cyclos_account(amount)
 
     @api.multi
     def addCyclosUser(self):
         for record in self:
+            backend_obj = self.env["res.partner.backend"]
             group = (
                 "particuliers" if record.company_type == "person" else "professionnels"
             )
@@ -262,10 +194,10 @@ class ResPartner(models.Model):
                 ],
             }
             try:
-                res = record._cyclos_rest_call("POST", "/users", data=data)
+                res = backend_obj._cyclos_rest_call("POST", "/users", data=data)
             except requests.exceptions.HTTPError as e:
                 if e.response.status_code == 422:
-                    msg = self._build_cyclos_error_message(e)
+                    msg = backend_obj._build_cyclos_error_message(e)
                     if msg != "":
                         raise ValueError(
                             "Cyclos serveur complained about:\n%s" % "\n".join(msg),
@@ -276,8 +208,7 @@ class ResPartner(models.Model):
             data = json.loads(res.text)
             if data:
                 _logger.debug("data: %s" % data)
-                backend_obj = self.env["res.partner.backend"]
-                backend_obj.create(
+                backend = backend_obj.create(
                     {
                         "partner_id": record.id,
                         "cyclos_id": data.get("user")["id"]
@@ -285,127 +216,10 @@ class ResPartner(models.Model):
                         else "",
                         "cyclos_status": data.get("status", ""),
                         "name": "cyclos:%s" % data.get("user")["id"],
+                        "type": "cyclos",
                         "status": "to_confirm",
                         "cyclos_create_response": res.text,
                     }
                 )
             if record.company_type == "person":
-                res = record.validateCyclosUser()
-
-    @api.multi
-    def validateCyclosUser(self):
-        for record in self:
-            backend_data = record._cyclos_backend()
-            res = record._cyclos_rest_call(
-                "POST", "/%s/registration/validate" % backend_data.cyclos_id
-            )
-            _logger.debug("res: %s" % res.text)
-            data = json.loads(res.text)
-            if data.get("status", False) and data.get("status") == "active":
-                backend_data.write(
-                    {
-                        "cyclos_status": data.get("status", ""),
-                        "status": "active",
-                        "cyclos_create_response": res.text,
-                    }
-                )
-
-    @api.multi
-    def activeCyclosUser(self):
-        for record in self:
-            backend_data = record._cyclos_backend()
-            data = {"status": "active", "comment": "Disable by Odoo"}
-            res = record._cyclos_rest_call(
-                "POST", "/%s/status" % backend_data.cyclos_id, data=data
-            )
-            backend_data.write(
-                {
-                    "status": "active",
-                    "cyclos_status": "active",
-                }
-            )
-            _logger.debug("res: %s" % res)
-
-    @api.multi
-    def blockCyclosUser(self):
-        for record in self:
-            backend_data = record._cyclos_backend()
-            data = {"status": "blocked", "comment": "Blocked by Odoo"}
-            res = record._cyclos_rest_call(
-                "POST", "/%s/status" % backend_data.cyclos_id, data=data
-            )
-            backend_data.write(
-                {
-                    "status": "blocked",
-                    "cyclos_status": "Blocked",
-                }
-            )
-            _logger.debug("res: %s" % res)
-
-    @api.multi
-    def disableCyclosUser(self):
-        for record in self:
-            backend_data = record._cyclos_backend()
-            data = {"status": "disabled", "comment": "Disable by Odoo"}
-            res = record._cyclos_rest_call(
-                "POST", "/%s/status" % backend_data.cyclos_id, data=data
-            )
-            backend_data.write(
-                {
-                    "status": "inactive",
-                    "cyclos_status": "disabled",
-                }
-            )
-            _logger.debug("res: %s" % res)
-
-    def forceCyclosPassword(self, password):
-        for record in self:
-            backend_data = record._cyclos_backend()
-            # TODO: need to stock password type id from cyclos API and replace -4307382460900696903
-            data = {
-                "newPassword": password,
-                "checkConfirmation": True,
-                "newPasswordConfirmation": password,
-                "forceChange": False,
-            }
-            try:
-                record._cyclos_rest_call(
-                    "POST",
-                    "/%s/passwords/%s/change"
-                    % (backend_data.cyclos_id, "-4307382460900696903"),
-                    data=data,
-                )
-            except ValueError as e:
-                if (
-                    len(e.args) > 1
-                    and e.args[1].status_code == 422
-                    and "newPassword" in e.args[1].json().get("properties", [])
-                ):
-                    _logger.debug("Ignoring Cyclos NewPassword Error !")
-                else:
-                    raise
-
-    def createCyclosUserToken(self, api_login, api_password):
-        self.ensure_one()
-        for record in self:
-            res = record._cyclos_rest_call(
-                "POST",
-                "/auth/session",
-                data={"timeoutInSeconds": 90000},
-                api_login=api_login,
-                api_password=api_password,
-            )
-            _logger.debug("res TOKEN: %s" % res.text)
-            data = json.loads(res.text)
-            return data.get("sessionToken", False)
-
-    @api.multi
-    def removeCyclosUserToken(self, api_login, api_password):
-        for record in self:
-            res = record._cyclos_rest_call(
-                "DELETE",
-                "/auth/session",
-                api_login=api_login,
-                api_password=api_password,
-            )
-            _logger.debug("res: %s" % res.text)
+                backend.validateCyclosUser()
